requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isodate==0.6.0
pandas==1.0.1
pyparsing==2.4.6
pywikibot==3.0.20190722
rdflib==4.2.2
requests==2.22.0
six==1.13.0
SPARQLWrapper==1.8.5
urllib3==1.25.7
xlrd==1.2.0
numpy==1.17.4
boto3==1.10.18
nltk==3.4.5
paramiko==2.7.1
pycookiecheat==0.4.5
pylint==2.4.4
openpyxl==3.0.5
google-cloud-translate==3.0.1